requirements.txt
langchain
langchain-community
langchain-weaviate
langchain-core
langchain-huggingface
langchain_mistralai
langgraph
langchain-openai
duckduckgo-search
weaviate-client
python-dotenv
